packages/mcp-server/python/requirements.txt
openpyxl>=3.1.0
weasyprint>=60.0
Jinja2>=3.1.0
mistune>=3.0.0
PyYAML>=6.0
python-docx>=1.1
